backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic[email]==2.7.4
pydantic-settings==2.3.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.7
meilisearch==0.31.4
celery==5.4.0
httpx==0.27.0
supabase==2.5.1
storage3==0.7.7
websockets==12.0
python-dotenv==1.0.1
Pillow==10.3.0
